requirements.txt
astroid==3.3.4
asttokens==2.4.1
autosemver==0.5.5
blinker==1.8.2
certifi==2024.8.30
colorama==0.4.6
configobj==5.0.9
coverage==7.6.1
dill==0.3.8
dulwich==0.19.16
enum34==1.1.10
exceptiongroup==1.2.2
executing==2.1.0
Faker==30.1.0
icecream==2.1.3
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.1
platformdirs==4.3.6
pluggy==1.5.0
Pygments==2.18.0
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pylint-pytest==1.1.8
pytest==8.2.0
pytest-cov==5.0.0
pytest-pylint==0.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
tomli==2.0.1
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
workflow==2.1.6
pandas==2.2.1